requirements.txt
asgiref==3.11.1
cfgv==3.5.0
distlib==0.4.0
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.24.3
identify==2.6.16
nodeenv==1.10.0
platformdirs==4.9.2
pre_commit==4.5.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
virtualenv==20.38.0
